+++ internal/models/OTP.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type OTP struct {
-	ID        uint      `gorm:"primaryKey"`
-	UserID    uint      `gorm:"not null"`
+	ID        int64     `gorm:"primaryKey"`
+	UserID    int64     `gorm:"not null"`
 	OtpCode   string    `gorm:"not null"`
 	ExpiresAt time.Time `gorm:"not null"`
 }
